refactor: reorganize agent command into modular structure
- Refactor 564-line monolithic agent.ts into organized folder structure
- Add shared utilities (types.ts, url-utils.ts) for cross-command reuse
- Implement modular agent command with logical groupings:
  - actions/: CRUD operations (crud.ts) and lifecycle management (lifecycle.ts)
  - manage/: validation schemas with Zod and display utilities
- Add comprehensive TypeScript validation with Zod schemas
- Maintain full backward compatibility and all existing exports
- Update tsup.config.ts with new entry points for proper build
- All 12 existing tests pass, confirming functionality preserved
