+++ packages/cli/src/commands/agent/actions/index.ts
@@ -1,0 +1,2 @@
+export * from './crud';
+export * from './lifecycle';
