+++ packages/cli/src/commands/agent/manage/index.ts
@@ -1,0 +1,2 @@
+export * from './validation';
+export * from './display';
