+++ packages/cli/src/commands/agent/types.ts
@@ -1,0 +1,7 @@
+/**
+ * Interface representing the payload sent when starting an agent.
+ */
+export interface AgentStartPayload {
+  characterPath?: string;
+  characterJson?: Record<string, unknown>;
+}
