+++ packages/cli/src/commands/shared/url-utils.ts
@@ -1,0 +1,19 @@
+import type { OptionValues } from 'commander';
+
+/**
+ * Helper function to determine the agent runtime URL
+ */
+export function getAgentRuntimeUrl(opts: OptionValues): string {
+  return (
+    opts.remoteUrl?.replace(/\/$/, '') || // Use the flag if provided
+    process.env.AGENT_RUNTIME_URL?.replace(/\/$/, '') || // Fallback to env var
+    `http://localhost:${opts.port || process.env.SERVER_PORT || '3000'}` // Use port flag or env var, default to 3000
+  );
+}
+
+/**
+ * Helper function to get the agents base API URL
+ */
+export function getAgentsBaseUrl(opts: OptionValues): string {
+  return `${getAgentRuntimeUrl(opts)}/api/agents`;
+}
